feat: add sorting, display toolbar on no results, add contain: strict, fix minor bugs
Add sort based on "Best match" (uses fuzzy search), name, weight and height,
plus the latter three reversed.
contain: strict is for better performance, and because of that,
the height of a card has to be calculated manually.
Fix height displaying next to the label weight on cards.
Run prebuild on npm run dev.