feat: show detailed screen when clicking a pokemon card
Change card to button for better keyboard support.
Use Nuxt's $router instead of replaceState (it seems
to work better with Nuxt, naturally).
Listen to $route changes so that the website can react
to the back and forward buttons.
The detailed screen shows "Types", "Stats" and
everything which was already included on the card.